# swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Service Management API",
    "description": "Auto-generated Swagger docs for Service Management Backend",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://service-management-server.onrender.com/api"
    },
    {
      "url": "http://localhost:8000/api"
    }
  ],
  "paths": {
...
    "/api/po-evaluations/po/{requestId}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "requestId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "requestId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "weights": {
                    "example": "any"
                  },
                  "comment": {
                    "example": "any"
                  },
                  "recommendedOfferId": {
                    "example": "any"
                  },
                  "offers": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/po-evaluations/{requestId}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "requestId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "requestId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "weights": {
                    "example": "any"
                  },
                  "comment": {
                    "example": "any"
                  },
                  "recommendedOfferId": {
                    "example": "any"
                  },
                  "offers": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Use: Bearer <token> (optional in your current backend)"
      },
      "userRole": {
        "type": "apiKey",
        "in": "header",
        "name": "x-user-role",
        "description": "Required: PROJECT_MANAGER / RESOURCE_PLANNER / PROCUREMENT_OFFICER / SYSTEM_ADMIN"
      },
      "username": {
        "type": "apiKey",
        "in": "header",
        "name": "x-username",
        "description": "Required for PM actions and ownership checks"
      }
    }
  }
}